server/requirements.txt
fastapi
uvicorn
pydantic
python-multipart
google-generativeai
langchain<0.2.0
langchain-community<0.1.0
langchain-google-genai<1.0.0
pypdf
faiss-cpu
python-dotenv
